Add files via upload
We created an interface to predict the top 5 crops that can be grown in an agricultural land by using Data Science. Mainly we used html for creating the structure of the interface where we can fill our details and for connecting it to the data base and the web server we used Flask for both the front end and back and connection, Python libraries for data bases and as a whole created an interface where we can give our details and get the accurate result.